+++ scm-ui/ui-components/src/index.ts
@@ -75,6 +75,7 @@ export { default as ErrorBoundary } from "./ErrorBoundary";
 export { default as OverviewPageActions } from "./OverviewPageActions";
 export { default as CardColumnGroup } from "./CardColumnGroup";
 export { default as CardColumn } from "./CardColumn";
+export { default as CardColumnSmall } from "./CardColumnSmall";
 
 export { default as comparators } from "./comparators";