+++ scm-webapp/src/main/java/sonia/scm/plugin/DefaultPluginManager.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.plugin;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -31,7 +31,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sonia.scm.NotFoundException;
 import sonia.scm.event.ScmEventBus;
-import sonia.scm.lifecycle.RestartEvent;
 import sonia.scm.lifecycle.Restarter;
 import sonia.scm.version.Version;
 
@@ -52,7 +51,6 @@ import static sonia.scm.ScmConstraintViolationException.Builder.doThrow;
 //~--- JDK imports ------------------------------------------------------------
 
 /**
- *
  * @author Sebastian Sdorra
  */
 @Singleton
@@ -64,17 +62,19 @@ public class DefaultPluginManager implements PluginManager {
   private final PluginCenter center;
   private final PluginInstaller installer;
   private final Restarter restarter;
+  private final ScmEventBus eventBus;
 
   private final Collection<PendingPluginInstallation> pendingInstallQueue = new ArrayList<>();
   private final Collection<PendingPluginUninstallation> pendingUninstallQueue = new ArrayList<>();
   private final PluginDependencyTracker dependencyTracker = new PluginDependencyTracker();
 
   @Inject
-  public DefaultPluginManager(PluginLoader loader, PluginCenter center, PluginInstaller installer, Restarter restarter) {
+  public DefaultPluginManager(PluginLoader loader, PluginCenter center, PluginInstaller installer, Restarter restarter, ScmEventBus eventBus) {
     this.loader = loader;
     this.center = center;
     this.installer = installer;
     this.restarter = restarter;
+    this.eventBus = eventBus;
 
     this.computeInstallationDependencies();
   }
@@ -172,8 +172,10 @@ public class DefaultPluginManager implements PluginManager {
         PendingPluginInstallation pending = installer.install(plugin);
         dependencyTracker.addInstalled(plugin.getDescriptor());
         pendingInstallations.add(pending);
+        eventBus.post(new PluginEvent(PluginEvent.PluginEventType.INSTALLED, plugin));
       } catch (PluginInstallException ex) {
         cancelPending(pendingInstallations);
+        eventBus.post(new PluginEvent(PluginEvent.PluginEventType.INSTALLATION_FAILED, plugin));
         throw ex;
       }
     }
@@ -255,7 +257,7 @@ public class DefaultPluginManager implements PluginManager {
 
       Set<String> dependencies = plugin.getDescriptor().getDependencies();
       if (dependencies != null) {
-        for (String dependency: dependencies){
+        for (String dependency : dependencies) {
           collectPluginsToInstall(plugins, dependency, false);
         }
       }